+++ wrappers/src/object_schema_cs.cpp
@@ -21,6 +21,7 @@
 #include "error_handling.hpp"
 #include "realm_export_decls.hpp"
 #include "object-store/src/shared_realm.hpp"
+#include "object-store/src/property.hpp"
 
 using namespace realm;