+++ wrappers/src/schema_cs.cpp
@@ -22,6 +22,7 @@
 #include "realm_export_decls.hpp"
 #include "object-store/src/shared_realm.hpp"
 #include "object-store/src/schema.hpp"
+#include "object-store/src/property.hpp"
 
 using namespace realm;